requirements.txt
numpy
pillow
opencv-python
scipy
google-genai
openai
moviepy
ffmpeg-python
pydub
librosa
soundfile
kokoro-tts
pytesseract
requests
python-dotenv
pyyaml
pandas
jsonschema
scikit-image
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
jupyter
ipython